requirements.txt
asgiref==3.7.2
channels==4.0.0
channels-redis==4.1.0
crispy-bootstrap5==2023.10
daphne==4.0.0
Django==4.2.7
django-allauth==0.59.0
django-channels==0.7.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
Pillow==10.1.0
redis==5.0.1
